+++ src/config.ts
@@ -13,7 +13,6 @@ import { createAPIClient } from './client';
  * When executing the integration in a development environment, these values may
  * be provided in a `.env` file with environment variables. For example:
  *
- * - `CLIENT_ID=123` becomes `instance.config.clientId = '123'`
  * - `CLIENT_SECRET=abc` becomes `instance.config.clientSecret = 'abc'`
  *
  * Environment variables are NOT used when the integration is executing in a
@@ -21,10 +20,11 @@ import { createAPIClient } from './client';
  * `instance.config` in a UI.
  */
 export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
-  clientId: {
-    type: 'string',
-  },
   clientSecret: {
+    type: 'string',
+    mask: true,
+  },
+  clientEmail: {
     type: 'string',
     mask: true,
   },
@@ -36,14 +36,14 @@ export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
  */
 export interface IntegrationConfig extends IntegrationInstanceConfig {
   /**
-   * The provider API client ID used to authenticate requests.
-   */
-  clientId: string;
-
-  /**
    * The provider API client secret used to authenticate requests.
    */
   clientSecret: string;
+
+  /**
+   * The client email at the provider.
+   */
+  clientEmail: string;
 }
 
 export async function validateInvocation(
@@ -51,9 +51,9 @@ export async function validateInvocation(
 ) {
   const { config } = context.instance;
 
-  if (!config.clientId || !config.clientSecret) {
+  if (!config.clientSecret || !config.clientEmail) {
     throw new IntegrationValidationError(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires all of {clientSecret, clientEmail}',
     );
   }