+++ src/steps/access/converter.ts
@@ -7,9 +7,9 @@ import {
 } from '@jupiterone/integration-sdk-core';
 
 import { Entities } from '../constants';
-import { AcmeGroup, AcmeUser } from '../../types';
+import { OrcaGroup, OrcaRole, OrcaUser } from '../../types';
 
-export function createUserEntity(user: AcmeUser): Entity {
+export function createUserEntity(user: OrcaUser): Entity {
   return createIntegrationEntity({
     entityData: {
       source: user,
@@ -17,18 +17,20 @@ export function createUserEntity(user: AcmeUser): Entity {
         _type: Entities.USER._type,
         _class: Entities.USER._class,
         _key: user.id,
-        username: 'testusername',
-        email: 'test@test.com',
-        active: true, // this is a required property
+        username: user.email,
+        email: user.email,
+        name: `${user.first_name} ${user.last_name}`,
+        active: true,
         // This is a custom property that is not a part of the data model class
         // hierarchy. See: https://github.com/JupiterOne/data-model/blob/master/src/schemas/User.json
-        firstName: 'John',
+        firstName: user.first_name,
+        lastName: user.last_name,
       },
     },
   });
 }
 
-export function createGroupEntity(group: AcmeGroup): Entity {
+export function createGroupEntity(group: OrcaGroup): Entity {
   return createIntegrationEntity({
     entityData: {
       source: group,
@@ -36,10 +38,21 @@ export function createGroupEntity(group: AcmeGroup): Entity {
         _type: Entities.GROUP._type,
         _class: Entities.GROUP._class,
         _key: group.id,
-        email: 'testgroup@test.com',
-        // This is a custom property that is not a part of the data model class
-        // hierarchy. See: https://github.com/JupiterOne/data-model/blob/master/src/schemas/UserGroup.json
-        logoLink: 'https://test.com/logo.png',
+        name: group.name,
+      },
+    },
+  });
+}
+
+export function createRoleEntity(role: OrcaRole): Entity {
+  return createIntegrationEntity({
+    entityData: {
+      source: role,
+      assign: {
+        _type: Entities.ROLE._type,
+        _class: Entities.ROLE._class,
+        _key: role.id,
+        name: role.name,
       },
     },
   });
@@ -74,5 +87,16 @@ export function createGroupUserRelationship(
     _class: RelationshipClass.HAS,
     from: group,
     to: user,
+  });
+}
+
+export function createUserRoleRelationship(
+  user: Entity,
+  role: Entity,
+): Relationship {
+  return createDirectRelationship({
+    _class: RelationshipClass.ASSIGNED,
+    from: user,
+    to: role,
   });
 }