+++ src/steps/access/index.ts
@@ -8,7 +8,7 @@ import {
 
 import { createAPIClient } from '../../client';
 import { IntegrationConfig } from '../../config';
-import { AcmeGroup } from '../../types';
+import { OrcaGroup } from '../../types';
 import { ACCOUNT_ENTITY_KEY } from '../account';
 import { Entities, Steps, Relationships } from '../constants';
 import {
@@ -16,7 +16,9 @@ import {
   createAccountUserRelationship,
   createGroupEntity,
   createGroupUserRelationship,
+  createRoleEntity,
   createUserEntity,
+  createUserRoleRelationship,
 } from './converter';
 
 export async function fetchUsers({
@@ -29,6 +31,7 @@ export async function fetchUsers({
 
   await apiClient.iterateUsers(async (user) => {
     const userEntity = await jobState.addEntity(createUserEntity(user));
+
     await jobState.addRelationship(
       createAccountUserRelationship(accountEntity, userEntity),
     );
@@ -45,9 +48,21 @@ export async function fetchGroups({
 
   await apiClient.iterateGroups(async (group) => {
     const groupEntity = await jobState.addEntity(createGroupEntity(group));
+
     await jobState.addRelationship(
       createAccountGroupRelationship(accountEntity, groupEntity),
     );
+  });
+}
+
+export async function fetchRoles({
+  instance,
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const apiClient = createAPIClient(instance.config);
+
+  await apiClient.iterateRoles(async (role) => {
+    await jobState.addEntity(createRoleEntity(role));
   });
 }
 
@@ -58,7 +73,7 @@ export async function buildGroupUserRelationships({
   await jobState.iterateEntities(
     { _type: Entities.GROUP._type },
     async (groupEntity) => {
-      const group = getRawData<AcmeGroup>(groupEntity);
+      const group = getRawData<OrcaGroup>(groupEntity);
 
       if (!group) {
         logger.warn(
@@ -85,6 +100,33 @@ export async function buildGroupUserRelationships({
   );
 }
 
+export async function buildUserRoleRelationships({
+  jobState,
+  instance,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const apiClient = createAPIClient(instance.config);
+
+  await apiClient.iterateAccessUsers(async (userWithRole) => {
+    const { user, role } = userWithRole;
+
+    const userEntity = await jobState.findEntity(user.id);
+
+    if (!userEntity) {
+      return;
+    }
+
+    const roleEntity = await jobState.findEntity(role.id);
+
+    if (!roleEntity) {
+      return;
+    }
+
+    await jobState.addRelationship(
+      createUserRoleRelationship(userEntity, roleEntity),
+    );
+  });
+}
+
 export const accessSteps: IntegrationStep<IntegrationConfig>[] = [
   {
     id: Steps.USERS,
@@ -103,11 +145,27 @@ export const accessSteps: IntegrationStep<IntegrationConfig>[] = [
     executionHandler: fetchGroups,
   },
   {
+    id: Steps.ROLES,
+    name: 'Fetch Roles',
+    entities: [Entities.ROLE],
+    relationships: [Relationships.USER_ASSIGNED_ROLE],
+    dependsOn: [Steps.ACCOUNT],
+    executionHandler: fetchRoles,
+  },
+  {
     id: Steps.GROUP_USER_RELATIONSHIPS,
     name: 'Build Group -> User Relationships',
     entities: [],
     relationships: [Relationships.GROUP_HAS_USER],
     dependsOn: [Steps.GROUPS, Steps.USERS],
     executionHandler: buildGroupUserRelationships,
+  },
+  {
+    id: Steps.USER_ROLE_RELATIONSHIPS,
+    name: 'Build User -> Role Relationships',
+    entities: [],
+    relationships: [Relationships.USER_ASSIGNED_ROLE],
+    dependsOn: [Steps.USERS, Steps.ROLES],
+    executionHandler: buildUserRoleRelationships,
   },
 ];