+++ src/types.ts
@@ -1,25 +1,86 @@
-// Providers often supply types with their API libraries.
+export interface OrcaUser {
+  id: string;
+  email: string;
+  first_name: string;
+  last_name: string;
+}
 
-export interface AcmeUser {
+export interface OrcaGroup {
+  id: string;
+  name: string;
+  description?: string;
+  sso_group: boolean;
+  users: { id: string }[];
+}
+
+export interface OrcaRole {
   id: string;
   name: string;
 }
 
-export interface AcmeGroup {
-  id: string;
-  name: string;
-  users?: Pick<AcmeUser, 'id'>[];
+export interface OrcaUserWithRole {
+  user: OrcaUser;
+  role: OrcaRole;
 }
 
-// Those can be useful to a degree, but often they're just full of optional
-// values. Understanding the response data may be more reliably accomplished by
-// reviewing the API response recordings produced by testing the wrapper client
-// (./client.ts). However, when there are no types provided, it is necessary to define
-// opaque types for each resource, to communicate the records that are expected
-// to come from an endpoint and are provided to iterating functions.
+// /api/organization/users
+export interface OrcaOrganizationUsersResponse {
+  status: string;
+  data: {
+    name: string;
+    users: {
+      user_id: string;
+      email: string;
+      first: string;
+      last: string;
+    }[];
+  };
+}
 
-/*
-import { Opaque } from 'type-fest';
-export type AcmeUser = Opaque<any, 'AcmeUser'>;
-export type AcmeGroup = Opaque<any, 'AcmeGroup'>;
-*/
+// /api/rbac/access/user
+export interface OrcaAccessUsersResponse {
+  status: string;
+  data: {
+    id: string;
+    user: OrcaUser;
+    role: OrcaRole;
+  }[];
+}
+
+// /api/rbac/group
+export interface OrcaGroupsResponse {
+  status: string;
+  data: {
+    groups: Omit<OrcaGroup, 'users'>[];
+  };
+}
+
+// /api/rbac/group/<id>
+export interface OrcaGroupResponse {
+  status: string;
+  data: {
+    group: string;
+    description?: string;
+    all_users: boolean;
+    users: {
+      id: string;
+    }[];
+  };
+}
+
+// /api/rbac/role
+export interface OrcaRolesResponse {
+  status: string;
+  data: OrcaRole[];
+}
+
+// /api/user/session
+export interface OrcaUserSessionResponse {
+  status: string;
+  role: string;
+  need_to_sign: boolean;
+  jwt: {
+    refresh: string;
+    access: string;
+  };
+}