+++ CHANGELOG.md
@@ -7,3 +7,15 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+Initial Orca Integration.
+
+- Ingest new entity `orca_account`
+- Ingest new entity `orca_user`
+- Ingest new entity `orca_group`
+
+- Build new relationship `orca_account_has_user`
+- Build new relationship `orca_account_has_group`
+- Build new relationship `orca_group_has_user`